Combine read_vr scripts
